Configure source mapping in development
The builtin PHP webserver was looking for existing static files (js, css)
under the `/src/assets` path, but relative to the public/ folder.
Obviously, these files don't exist (we don't want to give direct access
to the src/ folder!)

There were various options, but the lightest one was to take advantage
of the "router" option of the builtin server. It allows to not add a new
Nginx container (that would do the job more properly, but it's good
enough for the moment).