fix: Fix the locale form redirection
The locale form is aimed at redirect to the previous page. It needs the
current action pointer to do that correctly. It could have been set with
the current system of the canonical url, but I want a simpler system.

It also means I'll be able to use the new variable to declare the
canonical URL.