tec: Change assets folder destination
I want to bundle production assets for each release. The problem is I
need a different folder for dev and prod assets and I cannot rely
anymore on the single `static/` folder. So I created a new `url_asset()`
function which is pretty similar except it's looking in `dev_assets/` or
`assets/` folder depending on the current environment.