tec: Set error views only in web Application
The errors were handled by the cli Application class as well, but were
returned as HTML and so were unreadable in the console. Also, they were
badly handled because the views are using the "home" route name, which
is not initialized in the cli Application. An error was raised because
this route doesn't exist.