fix: Initialize the session from a custom cookie
The sessions were previously dropped due to the PHP session garbage
collector which dropped the value of the current_session_token variable.

The session cookie expiration is set back to 0 (i.e. until the browser is
closed) and a new flusio_session_token cookie is introduced to store the
session token. If the current user is not set yet, we try to initialize
it with this token.