fix: Fix title encoding on some websites
DOMDocument seems to try to guess the encoding, but was sometimes
failing. This patch tries to convert UTF-8 chars to HTML entities first.
This should cover a large part of websites, but is not perfect.

It could be improved later with encoding from the Content-Type header
and/or the encoding from the HTML charset meta tag. I didn't find any
website which doesn't work with this configuration though.

Reference: https://www.php.net/manual/domdocument.loadhtml.php#74777