fix: Fix margins of card__details
This is a regression from 90fa6ab. The .card__details class was used in
two different contexts and was designed differently in links.css and
collections.css.

Ideally the CSS class should be named differently, depending on the
context, but I wasn't able to find a better name in just few minutes so
I chose the easy solution.