fix: Fix NewsPicker duration test
Because of the test data, we could have a total reading time less than
the minimum duration time. For instance, for a duration of 30 min, the
minimum duration is 27 and maximum is 36. If the algo already got for 26
minutes of reading, the only available links must be less or equal to 10
minutes. If no such link exists, the total reading time would stay at
26, which is less than 27.

By reducing the links reading time, we make sure to always have a link
available.