tec: Fix the locale in User factory
The locale was badly generated because we were returning the full name
of the locale (i.e. "English" or "Français"). Because it was invalid,
the application was always set to english by default. Fixing this with
`array_keys` would return `en_GB` or `fr_FR`, which would break actual
tests which check for the presence of specific sentences.

To avoid this problem, we always return `en_GB`.