imp: Change cursor on button hover to pointer
I don't have a strong opinion about which kind of cursor to use. Some UX
designers have good arguments to keep the default behaviour. In the mean
time, my buttons don't always look like buttons and a pointer could
better convey the meaning of an action. Also, a LOT of well-known
websites and applications use the pointer on buttons.