imp: Hide "remove from news definitively" option
Its presence was confusing because "mark as read" has actually the same
effect. The difference is only visible in database (i.e. `is_read` vs.
`is_removed`), in anticipation of a feature to search for read articles.
The option was initially added to allow users to remove an article from
news without reading it.