fix: Fix modal undefined content
By quickly reopening the modal after closing it, it was possible to
have an "undefined" content. It was due to the small delay after the
closing which empties the modal body. I reduced the delay so it should
be impossible to reopen the modal before the delay. I also fixed the
call to properly empty the body.