fix: Don't redirect when discover page is empty
When the discover page was empty, "total pages" was equal to 0. Because
"current page" was higher to total pages (i.e. 1 > 0), it was set to the
value of "total pages" (i.e. 0).

However, in the controller, we tested that "page" param (1 by default)
was equal to the value of "current page". Because of the difference, we
returned a redirection, with "page" param set to "current page", i.e. 0.

Once redirected, because page param was less than 1, "current page" was
set to 1. The two values were still different, so we redirected the user
by setting the "page" param set to "current page", i.e. 1.

And so the user was traped in an infinite loop.

The reason was that we considered the current page of an empty
pagination should always be 1. However, "total pages" had a possible
value of 0, which didn't make sense.