fix: Make sure to initialize locale with default
The CLI didn’t necessarily initialize the default locale. Calling
`currentLocale()` would return `C` in this situation. Now we check the
locale is initialized when calling this method, and we set the default
locale during the initialization.

Running the tests against the `tests/cli/UsersTest.php` file only were
failing because of that (users didn’t have a valid locale on creation).