tec: Change the icon system
On slow networks, the interface was unusable while the icons were not
loaded. Because they were loaded asynchronously, they appeared long
after the HTML.

Now, the SVG icons are directly integrated in the HTML except very few
exceptions which should not be problematic. It also removes a lot of
code and a couple of hacks <3

Reference: https://github.com/flusio/flusio/issues/227