fix: Don't decode `+` during URL sanitization
I thought the `SpiderBits\Url::sanitize` method was idempotent but it
wasn't. Consequently, importations were failing for URLs containing a
`%2B` character because flusio wasn't able to find them in database.

`%2B` was first decoded in `+`. Later in the function, it re-encoded the
URL with `rawurlencode` which doesn't encode the `+` character. However,
by passing the result back to the method, it was first decoding the
string with `urldecode` which consider `+` to be a space. Then,
`rawurlencode` was encoding this space with `%20` (and not with a `+` as
`urlencode` would do).

The problem was obviously that it decoded and encoded the URLs with two
different standards.

Example:

```php
// Before:
$url1 = \SpiderBits\Url::sanitize('https://host/?query=foo%2Bbar');
$url2 = \SpiderBits\Url::sanitize($url1);
// $url1 and $url2 are different:
// https://host/?query=foo+bar != https://host/?query=foo%20bar

// After:
$url1 = \SpiderBits\Url::sanitize('https://host/?query=foo%2Bbar');
$url2 = \SpiderBits\Url::sanitize($url1);
// $url1 and $url2 are identical:
// https://host/?query=foo+bar == https://host/?query=foo+bar
```

References:

- https://en.wikipedia.org/wiki/Percent-encoding
- https://www.php.net/manual/function.rawurlencode.php
- https://www.php.net/manual/function.rawurldecode.php
- https://www.php.net/manual/function.urlencode.php
- https://www.php.net/manual/function.urldecode.php
- http://www.faqs.org/rfcs/rfc3986.html