sec: Reset sessions/reset token on password change
Considering an attacker who got a valid reset password link for a user,
he would be able to use it even after the user changed its email or
password himself.

The attack was mitigated by the short validity duration of the link
(1h), but it's a bit stronger now. Also, other sessions are deleted at
the same time, making sure a potential attacker is disconnected.