Sync feeds per batch of 25 each 15 seconds
The previous version tried to sync all feeds in 1 hour, which was
impossible with the number of feeds increasing.

Also, it was full bugged because, to calculate the number of feeds to
sync, we excluded the feeds that have been synced within the last hour
(instead of considering the total number of feeds).

This new version is way simpler to understand and should enable us to
sync feeds with multiple jobs in parallel later.