Remove FeedsFetcher job
This was disturbing to have two different jobs to handle (almost) the
same work. This makes things easier to understand.

The impact is that feeds that just have been imported from OPML will
take a bit more time to synchronize. Hopefully, this impact will be
reduced when the performance will increase.