Fetch feeds with two different strategies
With the last changes, feeds to fetch were selected randomly. It
means some feeds might never be selected with a high number of feeds.

This commit improves performance of the initial strategy by ignoring
inactive feeds, and add a second strategy to make sure to sync the
oldest feeds.

I’m not sure yet if the strategy selection ratio is right, it has to be
tested in production.