Allow several IPs to request Youtube
Youtube is blocking flusio if we make more than 1 request per minute.
However, if we use different IPs, there's a chance it will not detect
us. It means we can increase the rate limit to 2 or 3 requests per
minute (depending on the number of IPs are configured on the server). It
is not a lot but still 2 or 3x quicker.