Don't list all links during feeds sync
On app.flus.fr, there are almost 800,000 links associated to the
technical user. There are also 3 workers working in parallel on feeds
sync. In consequence, we can have 3 * 800000 = 2,400,000 ids loaded in
memory at the same time, and there are always more links in database!

With this solution, we only list the link ids of the current feed. The
bigger feeds contain 21,328, 11,822 and 9,732 links today, so a maximum
of ~43,000 ids loaded at the same time.

The counterpart is that we make an additional call to search the link in
the database each time the URL is not associated to the current
collection. This should not be problematic for existing feeds (most of
the links should already be in DB), and new feeds don’t stay new very
long.