Synchronize feeds that have changed only
A lot of feeds don’t change between two synchronizations. I don’t use
the information of "last update" from the feed because I don't trust it,
I prefer to calculate a hash which is more reliable, with the "risk" of
false positives (e.g. feed changing at each call). The risk is low
because the hash is not calculated from the raw content, but from the
parsed content.