Move response headers declaration in Application
This allows me to test the headers. Turbolinks-Location header can’t be
moved, or I would have to rebuild the `$http_uri` value, but I prefer to
use the real value. Hopefully, the next version of Turbolinks (i.e.
Turbo Drive) doesn’t require this header and I should be able to remove
it the day I’ll update the library.