Refactor CSRF mechanism
Two things bothered me with the previous way of managing CSRF tokens.

First was that I needed to instantiate a CSRF object each time I needed
to validate a token. I always felt like I was forgetting something, or
like I was doing a mistake in the syntax. In consequence, I always
copy-pasted the code from another method, which is fine but it is a
proof the developer experience was not good.

The second things was the `csrf_token()` view function. It was all
magical, with the generation of the token in background. It is in fact a
rest of time when the method `declareDefaultVariables()` didn’t exist. I
can now use this mechanism to remove the magic!

I really appreciate the new mechanism :)