Simplify generation of canonical URL
I previously tried to build the canonical URL with
`current_action_pointer`, which didn't work for URLs with parameters.
This is why I added then the `canonical` parameter to the views. But
then, it became hard to understand the difference between these two
parameters.

This commit simplifies the mechanism by removing the
`current_action_pointer` parameter and by making `canonical` required.