Refactor code of files upload
The code to test the files upload was very hacky and not easy to
understand. Also I don’t like the PHP API for files (why is there an
error_status when there is no error?)

I abstracted the file management behind a new Minz class (i.e. File) in
order to facilitate the code and the tests.