Simplify the CLI commands format
Mimic the Web request format is smart from a technical point of view,
but can be disturbing for an administrator who just want to use the CLI.

Also, I was facing some limitations from the `getopt()` function that
led me to propose a weird interface (i.e. `-pKEY=VALUE`).

The new solution breaks the old interface, but offers a better
experience for newcomers.