tec: Select links to fetch based on fetched_code
Using fetched_error was an error because the content of this column can
be really big. I was unable to add an index to this column in production
for instance.

This request is one of the most executed and took an average of 290ms to
execute. With these changes, it should be way more efficient.