sec: Duplicate identical URL for different feeds
By taking advantage of the uniquess of the links.url column, an attacker
was able to create a feed with entries pointing to external URLs. Then,
using the mechanism of the feed_entry_id update, he would be able to
change the URL back to its own website. Because the attacker and
external feeds shared the same link in DB, the external feed would have
an entry pointing to the attacker website too.

The disadvantage of this solution is that a normal user *can* duplicate
URLs. But it's theorical because I already check for this in the code.
The DB constraint was only an additional security.

I removed the bulk test because I can't test it anymore (not null
constraints fail either way).