imp: Handle different link publication dates
A link can be added to a collection at a time "T", and added to another
collection at a time "T+1". Until now, the only date we saved was "T".
In consequence, it was possible for a freshly added link to be burried
at the end of a collection, just because it has been added in another
collection long time ago.

Handling different dates has several implications such as changing the
date we display, the links order column, how Pocket and feeds
publication dates are saved and how we handle publication dates in the
news. I didn't expect so many changes in this commit and I expect more
to come as long as I detect impacts… but it's a long waited change and I
should not postpone it more.