Refactor use of LinksToCollections
LinksToCollections DAO model was painful to use. I now use a Model with
a DaoConnector to facilitate its manipulation. I also added a bunch of
methods shortcut so make the controllers code easier to read.

This commit also solves a small issue introduced by the previous
refactoring. When updating the collections of a link, it was removed
from the news and read list (because those collections were not in the
collection_ids array).