tec: Alter version in development and test environments
As the version will be used during exportation, this can be helpful in
the future to know if exported data have been generated by an unstable
version of flusio or not.

It can also help adminsys to know if the flusio instance which fetches
their servers is running in production or not (and to be a bit more
comprehensive if it's not :)).