Provide a DataExporter service
I created two new Atom files for collections and links to be more
flexible and to export such things as collections privacy and groups.

It was possible to use the `feed.atom.xml` files, but I prefer to make
it clear that they are not used in the same context.