tec: Reorganize media files
Some media folders contain more than 1 million of files in production…
listing the files in these folders is very slow and performance are
pretty bad.

The media folders are now splitted in subdirectories (3 levels of
depth). Each subdirectory is named according to the first characters of
the filename. For instance, the file `abcdefghijklm.png` can be found
under the subpath `abc/def/ghi/`.