tec: Delay groups loading on collections page
The performance was slow with a lot of collections and groups. I wasn't
sure how to tweak SQL requests to improve their performance.

Now, Turbo Frames loads the groups after the initial page load. I'm not
sure it is a good idea (e.g. scroll position is lost on page reload),
but I hope the collections page will be a bit more enjoyable to be
loaded.