tec: Close DB connections in jobs workers
I observed performance issues with the database on the server. I suspect
the workers to be the cause since they never close the connection to
the database. I’m not sure if it’s better to close/reopen connections at
each loop, but I’ll try to check the impact.