Remove until option of NewsPicker
I’ll add a new option to restrict links on a date-basis *per* feed, so I
need to move the logic of the `until` option. Also, the until option is
only used for the followed collections so I can remove the option
safely. I may want to bring it back in the future, but I'll see how I
design things at this moment.

I also changed a redundant test into a test that was missing.