tec: Clean old unused feeds and links
These data tend to accumulate for nothing in the database. It can be due
to users unfollowing a feed or deleting their accounts. There are more
than 3000 unfollowed feeds today in production, and almost the same of
links in no collections (but it will probably be more after the feeds
will be deleted).

I decided to keep unstored links for normal users because they can have
commented them and keeping the link somewhere else.