tec: Unlock jobs, links and collections after an hour
It can happen that a job, a link or a collection stays locked after a DB
restart for instance. If it happens, it can take some time to figure it
out. During this time, the job, or the link/collection sync is not done.

A timeout of 1 hour should be largely enough for all the existing jobs
to finish.