imp: Decrease space took by news buttons
These buttons take a lot of space on mobile, and we were able to see
only the first without scrolling. The images are now on the left, which
removes almost 150px of height from each button.

I had to modify a bit the reading illustration so it takes less
horizontal space (it was a bit unbalanced otherwise).