imp: Display new feed errors in the modal
For now, responses to form submits redirect to a new page. However,
I’d like the modal forms errors to be displayed directly in the modal,
without redirecting to a new page.

Since the modal is handled with a Turbo Frame, it should not be too
complicated. Except that's impossible. I need to set the frame attribute
target="_top" in order to redirect after a success, but I would need the
opposite in case of an error.

There are some discussions to handle this use case directly with Turbo,
but until an official solution appears, I will use this hack which use
Turbo Stream to update the modal content. I don’t think it’s a perfect
solution, so I’m not going to generalize it in the application. The new
feed form is important to handle correctly since errors are common (e.g.
a feed cannot be found).

Reference: https://github.com/hotwired/turbo/issues/138#issuecomment-847699281