imp: Homogeneize login modal
When a non-logged visitor tries to perform an action that need to be
logged in (e.g. following a collection or adding a link to a
collection), a modal is displayed to the visitor asking him to login.

Before, the "following a collection" action was the only one which
needed this modal, but after 826b3b1682529bde26da591902cb675dc4b7c43e,
it's not true anymore. I also realized I was able to load the login page
for free in the modal (Turbo <3).

This commit removes the specific code in favour of the default login
page.