imp: Accept is_hidden when adding from public collection
When users add a link from a public collection, they should be able to
set the is_hidden info as well.

I’m not happy with this `mode` solution which is super repetitive, but
it works easily. I’ll probably want to refactor this code with the
"Obtentions" controller in the future.