fix: Fix performance issue on collections pages
On collections pages, the "read" status was calculated by resending one
SQL request per link to the database. For a full collection, that means
+30 requests. It's not a big deal with a small database, but I have more
than 4 millions of links in production. Loading a collection could take
more than 3 seconds.

This commit introduces a new "is_read" computed property in the Link
model, which is calculated during the listing with a more efficient
request (with a WITH statement, which acts like a temporary table). It
is a bit verbose, but fortunately the DAO file is not as big as it was
before so it stays readable.