misc: Refactor sorting methods
Each model shouldn't have to declare its own sorting method. It belongs
to the controller (or the view) to know by which property it wants to
sort the models.

Collections are no longer sorted by type (bookmarks having the
priority), but I updated the code in order to keep the same order in
practice.

I removed collections from the links/Messages controller because they
were unused in the view.