imp: Revert delaying groups loading on /collections
The interest was obvious from a performance perspective, but there were
some drawbacks. For instance the groups could take a lot of time to
load, without being noticed by the user. If she tried to ctrl+f the name
of such a group, it wouldn't show.

With the latest improvements on the DAO, performance should be better
than before. Also, this page will be dropped once the new navigation
will be enabled. If performance should be bad, it will not be for a long
time.

This change allows to remove an option from the listComputedByUserId
Collection DAO method and the $options parameter from
listComputedFollowedByUserId. It's always nice to drop some code :)

Reference: 2d575d22ad79b03a7688a055edfedafb55d5cc31