cli: Reset the locale after each job operation
I didn't understand why the demo user was reset with English locale
whereas the CLI_LOCALE was set to French. In fact, the locale is changed
during links synchronization with the support user locale, which is
English on the demo server. I could have changed the locale of the
support user, but I think it's better to force a reset after each job.