imp: Refactor and fix back navigation
Profiles feature (6c99f4f) introduced loop issues with the back
navigation. This was something I wanted to improve for a long time
without finding a correct solution.

My solution is to stack path history in an array and pop items when we
click on the back button. The difficulty was to handle the back and
forward navigation of the browser. I'm pretty sure there are still nasty
bugs, but I think I tested most of the possibility.

This solution also allows me to display the name of the previous page.