Fix InitNewDefaultCollections migration
The migration tests were failing because, at this point, the database
doesn't know about the new `autoload_modal` property. But because the
support user is initialized with properties declared in the model during
the tests (see the models\User::supportUser() call), it fails.

I ALWAYS MUST use the DAO directly in the migrations and avoid calling
"magic" methods. I also should explicit the database columns during
insertions without guessing them based on the declarations of the model.

To completely fix this test, I should have changed how the collections
are created, but this code is not executed during the tests, and the
migration is quite old so I don't think it's worth of my time.