misc: Migrate date formatting to IntlDateFormatter
The `strftime` function is deprecated in PHP 8.1 and will be removed in
PHP 9. I have time before getting there, but I wanted to fix that before
I forget.

I should have fixed the `_d` function in Minz, but it would break other
applications. Also, the current locale is not easily accessible to Minz
and it would require a different mechanism.

References:

- https://wiki.php.net/rfc/deprecations_php_8_1#strftime_and_gmstrftime
- https://www.php.net/manual/function.strftime
- https://php.watch/versions/8.1/strftime-gmstrftime-deprecated