imp: Improve accessibility of the skip nav button
Orca wasn’t announcing the "back" button after skipping the main
navigation. Now, it behaves strangely: it announces the back button
right after clicking on "skip to main content", but not after selecting
it with tab. A solution is to add a sr-only sentence (e.g. "Main
content") at the very beginning of <main> element, but I don't think
it's a better idea.

I’ll stay with the current solution (which works) until I find a better
one.