fix: Handle news batch actions redirections correctly
I didn't see there was a problem there. The collections/Read controller
expects a `from` parameter to redirect the user, but there was none.
So flusio was sending an empty redirection.

Firefox was handling this redirection by GETting the /read/later route.
Because this route doesn't exist, it was resulting in a 404 error and (I
think) Turbo Drive reloaded the page. Everything looked fine.

Chrome, however, didn't do anything on the empty redirection and so the
page wasn't reloaded.