Add full text search index to links table
I chose to use the "french" dictionnary because flusio is mainly used by
French people. In the future, a better solution would be to store the
index configuration in a different column. For now, it avoids me complex
code to handle the dictionnary.

Reference: https://www.postgresql.org/docs/current/textsearch-tables.html#TEXTSEARCH-TABLES-INDEX