Change the URL column used to find links
All the `links.url` column contains the scheme (http or https), which is
pretty bad for the index. The idea is, for now, to provide an additional
`url_lookup` column containing the URL without the scheme and index this
one instead of `url`.

This column is generated automatically with a PostgreSQL function so I
don't have to worry about it in the code.

The difficulty was to find every place where I was searching for the URL
and adapt the code. Performance can be pretty bad in production since I
removed the index from the `url` column.