fix: Fix changing collections of not owned links
Consider that a user A had a link with a given URL in its owned
collections. If he opened collections of a link with the same URL but
owned by user B, he would not see that the link was already in its
collections. Worse: submitting the form would delete the link from its
collections if they were not re-selected manually.