tec: Improve perf when displaying link collections
With the last sharing feature came the need to display when a URL was
already in a collection by another user (`link.url_lookup = 'URL' and
link.user_id != 'current user id'`). There was already an index on the
couple `user_id` / `url_lookup`, but it was ineffective in this
situation. This led to very poor performance in production.

Adding an index on just `url_lookup` solved the issue. I pushed the
solution to change also the `LEFT JOIN` to a `EXISTS` clause, which is
slightly more performant.