fix: Fix the strategy to sanitize URLs
The previous strategy was generating bad URLs, for instance:

```
https://example.com/?url=http%3A%2F%2Fexample.fr%2F%3Ffoo%3Dbar&spam=egg
```

was sanitized to:

```
https://example.com/?url=http://example.fr/?foo=bar&spam=egg
```

instead of:

```
https://example.com/?url=http%3A%2F%2Fexample.fr%2F%3Ffoo%3Dbar&spam=egg
```

The issue is that in the first case, we weren't able anymore to tell if
the `spam` parameter was associated to the base URL or to the `url`
value.

Now, each part of the URL is encoded separately.

`query` is re-encoded by parsing it, encoding each parameter one by one
and rebuilding the query. This way, we are sure that the `spam`
parameter is processed separately from the `url` parameter.

This fix changes the result of some tests, but they look okay and normal
considering the expected behaviour of the example given here.