fix: Fix the Url::percentRecodeQuery method
URLs can have several parameters with the same name (e.g. `?foo=bar&foo=baz`),
which are accessible in an array (see `Url::parseQuery()`).

However, the `percentRecodeQuery` method wasn't aware of this
possibility and was executing the `fullPercentDecode()` method directly
on the array instead of each part of the array.