fix: Fix tests in MediaTest failing randomly
The issue was that `$subdir_names_depth_1` included the dirs `.` and
`..` AND that the dirs were not sorted.

Locally, the dirs were always in this order: `., .., real_dir`. But on
the CI, the order was random, e.g. `., real_dir, ..`. The calculation of
the progress could be different from one run to another because it is
based on the index.

Reference: https://github.com/flusio/flusio/actions/runs/3117597151/jobs/5056358531