fix: Fix various outlines style
The "back" link outline was overflowing over the navigation bar and it
was covered by the section image if it existed.

In compact mode, the outline of links images was a square while the
images were rounded.

The outline of links titles and links comments was truncated in compact
mode.