fix: Try to fix a bug on links actions
Apparently, a bug happens on links actions on Vivaldi 5.4.2760.36 and
Android 9. I wasn't able to reproduces it on the same version of Vivaldi
and Android 8 though.

The bug: it seems that forms with `data-turbo-preserve-scroll` aren't
submitted, while the others are. The good news is there's very little
code to handle the "preserve scroll" behaviour, so the bug should be
easy to fix. I really expect the bug being `target` not available in the
`event` object in some circumstances. The bad news is that, if `target`
is `undefined`, the normal behaviour should be to submit the form
outside of Turbo (this is what happens when I change `target` for an
obviously `undefined` variable).

Another possibility is that `addEventListener` had to be called on
`document`. It was initially a mistake in my code, but it worked on all
browsers. Also, I think if it had to break something, it would have
break more than just the forms with `data-turbo-preserve-scroll`.

The last possibility (and the least expected) is that
`Turbo.navigator.currentVisit.scrolled` cannot be accessed and break the
code when set. I really don't think it's the problem because the code is
called after the form is submitted. Also, it should not be
platform-specific (i.e. it should be accessible, since it is on other
platforms).