tec: Force xdebug version in the Dockerfile
The latest version of xdebug doesn't support PHP 7.4 anymore. We need to
force the version to the latest version which support PHP 7.4.

The other solution would have been to upgrade to a higher PHP version,
but I want it to be the same as the production, which run on Debian.
I’ll update PHP when Debian 12 (Bookworm) will be released.

References:

- https://xdebug.org/docs/compat
- https://chrisshennan.com/blog/fixing-pecl-xdebug-requires-php-version-8-installed-version-7-4
- https://wiki.debian.org/PHP#Available_versions