fix: Compare LinksToCollections only on created_at
In FeedFetcher, we order the LinksToCollections being created, in order
to apply filters on the list. If two LinksToCollections had the same
created_at, we ordered then on the id. Except that the id is not set at
this step, since they are not created yet.

In PHP 8.2, accessing id failed and some feeds weren't synchronized
anymore.