imp: Autosave content of textareas
Textareas are used to enter long text. It's painful for the users when
they lose what they typed. Losing content can happen in different
situations, such as an unfortunate refresh, a browser crash, or, more
probably, if the user close the modal containing the textarea.

The content is stored in the localStorage when the user stops typing for
300 milliseconds.