fix: Sanitize URL queries containing spaces correctly
Spaces are encoded as `+` (or `%20`) in URLs. Unfortunately I decoded
this character as a simple `+` instead of a space. When I re-encoded the
query value, the `+` was then encoded as `%2B`, which was wrong as it
should have stayed as `+` (or `%20`).

Now, I we re-encode queries, the character `+` is correctly decoded to a
space, and then re-encoded to `+`.

Bug introduced in: c41afda21953f47257d88be0551e3f62e2f3b7a2