fix: Fix group headers positionning
The previous hack didn't work as expected for the collections and feeds
groups. After a certain number of groups, the header passed behind the
cards.

There is still an issue on mobile: if you open the actions menu of a
date group (not the first) and that you scroll up, the menu will appear
behind the headers. That's however a bit less problematic. Also, I plan
to change the popups behaviour by not fixing them to the screen.