fix: Generate correct links to tags in Atom feeds
In the feeds, the links to tags were in the form of `/links?q=#tag`.
Most of the aggregators are able to rebuild the correct link by using
the feed domain as the base URL. But it didn't work when the feeds were
generated with the `direct=true` parameter because the base URL wasn't
the one to Flus.

Now, the tags links are always generated in an absolute form to avoid
this issue (even in the interface because it was easier to patch).