fix: Fix the account validation mechanism
Some emails clients (e.g. Thunderbird) send an HTTP request to the URLs
before the user click on them in order to protect them against malicious
links. This request validated the account, but then when the users
arrived on the page, they were welcomed with a "The token is invalid"
error.

I've reworked the mechanism to have a separate page with a button to
click on. At the end, the code is way better and easier to maintain!