chore: Improve getting the collections and links to fetch
The previous system wasn't performant because of the random order.
Also, with few feeds or links to fetch, multiple parallel jobs would get
the same items to fetch at the same time. Because of an issue in the
lock mechanism, this would trigger multiple calls to the same URLs at the
same time.

The new serie system allows to improve the distribution of links and
collections to fetch across the different jobs.